Make integer constants 64bit clean
Interally Scheme integers and LLVM integer constants were represented
with scala.Int. This is a signed 32bit integer and cannot represent the
full signed 64bit range of our exact integers. Switch to scala.long and
add unit tests for both parsing and code generation.

+++ compiler/src/main/scala/llambda/AbstractSyntaxTree.scala
@@ -28,7 +28,7 @@ object FalseLiteral  extends BooleanLiteral(false)
 
 sealed abstract class NumberLiteral extends NonSymbolLeaf 
 
-case class IntegerLiteral(value : Int) extends NumberLiteral {
+case class IntegerLiteral(value : Long) extends NumberLiteral {
   override def toString = value.toString
 }
