+++ compiler/src/main/scala/llambda/codegen/LiveExactInteger.scala
@@ -4,7 +4,7 @@ import llambda.codegen.{boxedtype => bt}
 import llambda.codegen.llvmir._
 import llambda.nfi
 
-private class ConstantLiveExactInteger(module : IrModuleBuilder)(constantValue : Int) extends ConstantLiveValue(bt.BoxedExactInteger) {
+private class ConstantLiveExactInteger(module : IrModuleBuilder)(constantValue : Long) extends ConstantLiveValue(bt.BoxedExactInteger) {
   def genBoxedConstant() : IrConstant = {
     val boxedIntName = module.nameSource.allocate("schemeExactInteger")
 
@@ -59,7 +59,7 @@ private class UnboxedLiveExactInteger(unboxedValue : IrValue, nativeType : nfi.I
 }
 
 object LiveExactInteger {
-  def fromConstant(module : IrModuleBuilder)(value : Int) : ConstantLiveValue =
+  def fromConstant(module : IrModuleBuilder)(value : Long) : ConstantLiveValue =
     new ConstantLiveExactInteger(module)(value)
 
   def fromUnboxed(unboxedValue : IrValue, nativeType : nfi.IntType) : LiveValue =
